Created a Photon Based Tester
This tester uses the TPP Wireless I/O Board to stress test the LoRa sensor.  The relay on the Wireless I/O Board is closed and opened a pre-determined number of times.  The Hub logs the received messages to a Google Sheet using Particle.publish on the Hub, a Webhook in the Particle Cloud that communicates with a Google Apps Script that logs the sensor trips messages to a Google sheet.