Added support for address jumpers
Address jumpers are now processed in the code.  Changed setAdrress() argument type to unsigned int in all files.  Added pin definitions to globals.h file.

This code should work on the Photon 2 as well as on Arduino, but it has only been tested on Arduino.  These files need to be copied into the VS locations for Photon.