Test version 2.6.1 of Arduino
Added sending a reset test message at the end of setup() so that any resets during stress testing will be logged.  This is temporary code for stress testing the production sensor code on ATMega328 only.  No #defines or such are included for the Photon version.  This code is intended to run ONLY on an ATMega328 based sensor and only for stress testing purposes.